Rearange the e2e package:
* Move tests for the old UI from 'integrations' to 'legacy' (effectively omitting them from test runs)
* Move 'integrations/proto' -> 'integrations' (as they're the main tests now)
* Move 'helpers' -> 'helpers/legacy'
* Move 'helpers/proto' -> 'helpers'
