+++ apps/e2e/helpers/legacy/db.ts
@@ -1,0 +1,20 @@
+import type { Page } from "@playwright/test";
+
+import type { InventoryDatabaseInterface } from "@librocco/db";
+
+/**
+ * Returns the database handle from the db interface registered in the window object
+ * of the app. We can run `hadle.evaluate` to run queries/mutations against the database.
+ * @example
+ * ```ts
+ * const dbHandle = await getDbHandle(page);
+ *
+ * // Use the handle to create a warehouse in the db
+ * await dbHandle.evaluate(async (db) => {
+ *   await db.warehouse("foo-wh").create()
+ * })
+ * ```
+ */
+export function getDbHandle(page: Page) {
+	return page.evaluateHandle(() => window["_db"] as InventoryDatabaseInterface);
+}
