+++ apps/e2e/helpers/legacy/index.ts
@@ -1,0 +1,2 @@
+export { getDashboard } from "./dashboard";
+export { getDbHandle } from "./db";
