--- apps/e2e/helpers/proto/index.ts
@@ -1,3 +1,0 @@
-export { getDashboard } from "./dashboard";
-
-export { getDbHandle } from "@/helpers/db";
