+++ apps/e2e/helpers/types.ts
@@ -1,13 +1,17 @@
-import type { Locator } from "@playwright/test";
+/* eslint-disable @typescript-eslint/ban-types */
+import type { Locator, Page } from "@playwright/test";
 
-import type { NoteState, NoteTempState } from "@librocco/shared";
+import type { NoteState, NoteTempState, WebClientView } from "@librocco/shared";
 
 export type WaitForOpts = Parameters<Locator["waitFor"]>[0];
 export type GetByTextOpts = Parameters<Locator["getByText"]>[1];
 /** A type of display row property names, without 'warehouseId' as it's never displayed */
 export type TransactionRowField = keyof Omit<DisplayRow, "warehouseId">;
 
-export type ViewName = "inbound" | "outbound" | "stock";
+export type DashboardNode<T = {}> = T &
+	Locator & {
+		dashboard: () => DashboardInterface;
+	};
 
 interface Asserter<T> {
 	assert: (value: T, opts?: WaitForOpts) => Promise<void>;
@@ -35,24 +39,26 @@ export interface DisplayRow {
 }
 
 export interface DashboardInterface {
+	page(): Page;
 	waitFor: Locator["waitFor"];
 	nav(): MainNavInterface;
-	navigate(to: ViewName, opts?: WaitForOpts): Promise<void>;
-	view(name: ViewName): ViewInterface;
-	sidebar(): SidebarInterface;
+	navigate(to: WebClientView, opts?: WaitForOpts): Promise<void>;
+	view(name: WebClientView): ViewInterface;
 	content(): ContentInterface;
-	bookForm(): BookFormInterface;
+	// sidebar(): SidebarInterface;
+	disableNotifications: () => Promise<void>;
+	// bookForm(): BookFormInterface;
 }
 
 export interface NavInterface extends Locator {
 	link(label: string, opts?: { active?: boolean }): Locator;
 }
 
-export interface MainNavInterface extends Locator {
-	navigate(to: ViewName): Promise<void>;
-}
+export type MainNavInterface = DashboardNode<{
+	navigate(to: WebClientView): Promise<void>;
+}>;
 
-export type ViewInterface = Locator;
+export type ViewInterface = DashboardNode;
 
 export interface SidebarInterface extends Locator {
 	createWarehouse(opts?: WaitForOpts): Promise<void>;
@@ -69,21 +75,25 @@ export interface SideLinkGroupInterface extends Omit<SidebarInterface, "assertGr
 	assertClosed(): Promise<void>;
 }
 
-export interface ContentInterface extends Locator {
-	heading(title?: string, opts?: GetByTextOpts & WaitForOpts): ContentHeadingInterface;
-	updatedAt(opts?: WaitForOpts): Promise<Date>;
-	assertUpdatedAt(date: Date, opts?: WaitForOpts & { precision: number }): Promise<void>;
-	discount(): WarehouseDiscountInterface;
-	statePicker(): StatePickerInterface;
-	scanField(): ScanFieldInterface;
-	entries(view: ViewName): EntriesTableInterface;
-}
+export type ContentInterface = DashboardNode<{
+	header(): ContentHeaderInterface;
+	// updatedAt(opts?: WaitForOpts): Promise<Date>;
+	// assertUpdatedAt(date: Date, opts?: WaitForOpts & { precision: number }): Promise<void>;
+	// discount(): WarehouseDiscountInterface;
+	// statePicker(): StatePickerInterface;
+	// scanField(): ScanFieldInterface;
+	// entries(view: WebClientView): EntriesTableInterface;
+}>;
 
-export interface ContentHeadingInterface extends Locator {
-	getTitle(opts?: WaitForOpts): Promise<string>;
-	textInput(): Locator;
-	rename(newTitle: string, opts?: WaitForOpts): Promise<void>;
-}
+export type ContentHeaderInterface = DashboardNode<{
+	title: () => Asserter<string>;
+}>;
+
+// export interface ContentHeadingInterface extends Locator {
+// 	getTitle(opts?: WaitForOpts): Promise<string>;
+// 	textInput(): Locator;
+// 	rename(newTitle: string, opts?: WaitForOpts): Promise<void>;
+// }
 
 export interface StatePickerInterface extends Locator {
 	getState(opts?: WaitForOpts): Promise<NoteState | NoteTempState>;
