+++ apps/e2e/integration/constants.ts
@@ -1,0 +1,1 @@
+export const baseURL = process.env.CI ? "http://localhost:4173/preview/" : "http://localhost:5173/preview/";
