--- apps/e2e/integration/proto/constants.ts
@@ -1,1 +1,0 @@
-export const baseURL = process.env.CI ? "http://localhost:4173/preview/proto/" : "http://localhost:5173/preview/proto/";
