+++ apps/e2e/legacy/scan_flow.spec.ts
@@ -2,7 +2,7 @@ import { expect, test } from "@playwright/test";
 
 import { baseURL } from "../constants";
 
-import { getDashboard, getDbHandle } from "../helpers";
+import { getDashboard, getDbHandle } from "@/helpers/legacy";
 
 test.beforeEach(async ({ page }) => {
 	// Load the app
