+++ apps/e2e/legacy/stock_changes.spec.ts
@@ -3,7 +3,7 @@ import { expect, test } from "@playwright/test";
 
 import { baseURL } from "../constants";
 
-import { getDashboard, getDbHandle } from "../helpers";
+import { getDashboard, getDbHandle } from "@/helpers/legacy";
 
 test.beforeEach(async ({ page }) => {
 	// Load the app
